+++ tests/decompiler_tests.py
@@ -302,13 +302,22 @@ class DecompilerStartDecompilationTests(BaseServiceTests):
             params=AnyParamsWith(sel_decomp_decoding='only')
         )
 
-    def test_raw_endian_is_set_to_correct_value_when_given(self):
+    def test_endian_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            endian='little'
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(endian='little')
+        )
+
+    def test_raw_endian_is_still_supported_as_alias_for_endian(self):
         self.start_decompilation_with_any_input_file(
             raw_endian='little'
         )
 
         self.assert_post_request_was_sent_with(
-            params=AnyParamsWith(raw_endian='little')
+            params=AnyParamsWith(endian='little')
         )
 
     def test_raw_entry_point_is_set_to_correct_value_when_given(self):